Added the apache collections source files to simplify execution
This is required so that the TA/Prof can build and execute the code
without having to specify the full classpath (which is a pain).

Preferrably they should just use the make file and follow the README
but this makes it so they can just do a simple javac *.java, java Minimp ...
and it all "works".