FIXED: scoping, functions and if statements.
Scoping bugs were found since ASTDeclaration would not search for previous versions of the variable to use. Also, A major bug was fixed with if statements and else statements. These were not able to pass along the return value give a function return. Finally another major bug was found in the binding of parameter variables to the passed parameters. All of these have been fixed.