Added support to the test suite to generate ASM and execute it
The exemplar output files for the ASM code that is executed are in
the test/ directory, when the test suite is run each asm file generated
is stored in the asm/ directory and the output for the asm file executed
has the same name as the test file execpt with a ".out".

Any tests that have a "FAILED" in yellow failed due to differing output
or crashing when the ASM file was executed. Any "FAILED" in red are due
to crashes during parsing/interpreting (the actual compiler fails!).